Use ServiceMemory everywhere
